# INSTA-LITE/insta-lite-back/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  like.Like:
    properties:
      created_at:
        type: string
      postID:
        type: string
      reactionType:
        $ref: '#/definitions/like.ReactionType'
      updated_at:
        type: string
      userID:
        type: string
    type: object
  like.LikeRequest:
    properties:
      reactionType:
        $ref: '#/definitions/like.ReactionType'
    required:
    - reactionType
    type: object
  like.ReactionType:
    enum:
    - like
    - love
    - laugh
    type: string
    x-enum-varnames:
    - ReactionLike
    - ReactionLove
    - ReactionLaugh
...
  post.PaginatedPostResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/post.PostResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
    type: object
  post.Post:
    properties:
      commentsCount:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      likesCount:
        type: integer
      postID:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  post.PostCreateRequest:
    properties:
      content:
        type: string
      medias:
        items:
          $ref: '#/definitions/postmedia.PostMedia'
        type: array
    required:
    - content
    - medias
    type: object
  post.PostDeleteRequest:
    properties:
      postID:
        type: string
    required:
    - postID
    type: object
  post.PostResponse:
    properties:
      commentsCount:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      likesCount:
        type: integer
      media:
        items:
          $ref: '#/definitions/postmedia.PostMedia'
        type: array
      postID:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  post.PostUpdateRequest:
    properties:
      content:
        type: string
      postID:
        type: string
    required:
    - content
    - postID
    type: object
  postmedia.PostMedia:
    properties:
      createdAt:
        type: string
      mediaID:
        type: string
      postID:
        type: string
      type:
        enum:
        - image
        - video
        type: string
      url:
        type: string
    required:
    - type
    - url
    type: object
host: localhost:8080
info:
  contact:
    email: huguescodeur@gmail.com
    name: Hugues Codeur
  description: API REST pour une application de type Instagram (posts, likes, commentaires,
    follow).
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Insta-Lite API
  version: "1.0"
paths:
...
  /posts:
    delete:
      consumes:
      - application/json
      description: Supprime un post appartenant à l'utilisateur connecté
      parameters:
      - description: ID du post à supprimer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/post.PostDeleteRequest'
      responses:
        "204":
          description: Post supprimé
        "400":
          description: Format JSON invalide
          schema:
            type: string
        "404":
          description: Post introuvable
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Supprimer un post
      tags:
      - Posts
    get:
      description: Retourne les posts paginés par curseur (cursor-based pagination)
      parameters:
      - description: Curseur de pagination (opaque string)
        in: query
        name: cursor
        type: string
      - description: 'Nombre de résultats (défaut: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PaginatedPostResponse'
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Liste tous les posts (feed)
      tags:
      - Posts
    patch:
      consumes:
      - application/json
      description: Met à jour le contenu d'un post appartenant à l'utilisateur connecté
      parameters:
      - description: Données de mise à jour
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/post.PostUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.Post'
        "400":
          description: Format JSON invalide ou validation échouée
          schema:
            type: string
        "404":
          description: Post introuvable ou non autorisé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Modifier un post
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: Crée un nouveau post avec au moins un média
      parameters:
      - description: Contenu et médias du post
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/post.PostCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/post.PostResponse'
        "400":
          description: Format invalide ou aucun média
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Créer un post
      tags:
      - Posts
  /posts/{id}:
    get:
      description: Retourne les posts d'un utilisateur précis, paginés par curseur
      parameters:
      - description: UUID de l'utilisateur
        in: path
        name: id
        required: true
        type: string
      - description: Curseur de pagination
        in: query
        name: cursor
        type: string
      - description: 'Nombre de résultats (défaut: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PaginatedPostResponse'
        "400":
          description: ID incorrect
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Posts d'un utilisateur
      tags:
      - Posts
...
  /posts/{postID}/likes:
    delete:
      description: Retire la réaction de l'utilisateur connecté sur un post
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      responses:
        "204":
          description: Like supprimé
        "400":
          description: ID invalide
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Supprimer un like sur un post
      tags:
      - Post Likes
    get:
      description: Retourne tous les likes d'un post
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/like.Like'
            type: array
        "400":
          description: ID invalide
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Likes d'un post
      tags:
      - Post Likes
    patch:
      consumes:
      - application/json
      description: Change le type de réaction (like/love/laugh) de l'utilisateur connecté
        sur un post
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      - description: Nouvelle réaction
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/like.LikeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/like.Like'
        "400":
          description: ID ou type de réaction invalide
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Modifier une réaction sur un post
      tags:
      - Post Likes
    post:
      consumes:
      - application/json
      description: Ajoute une réaction (like/love/laugh) à un post
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      - description: Type de réaction
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/like.LikeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/like.Like'
        "400":
          description: Type de réaction invalide
          schema:
            type: string
        "404":
          description: Post non trouvé
          schema:
            type: string
        "409":
          description: Déjà liké
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Liker un post
      tags:
      - Post Likes
  /posts/{postID}/likes/me:
    get:
      description: Retourne le like de l'utilisateur connecté sur un post
      parameters:
      - description: UUID du post
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/like.Like'
        "400":
          description: ID invalide
          schema:
            type: string
        "404":
          description: Like non trouvé
          schema:
            type: string
        "500":
          description: Erreur interne
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Mon like sur un post
      tags:
      - Post Likes
...
securityDefinitions:
  BearerAuth:
    description: 'Entrez le token JWT avec le préfixe "Bearer ". Exemple: "Bearer
      eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"